Feat: Start the authentication

+++ base/views.py
@@ -1,7 +1,44 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
+
+def loginPage(request):
+    """ Login page view """
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist.")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username / Password does not match")
+
+    context = {}
+    return render(request, "base/login_register.html", context)
+
+def logoutUser(request):
+    """ Logout function for the user """
+
+    logout(request)
+    return redirect('home')
 
 def home(request):
     """ Home page """
@@ -31,6 +68,7 @@ def room(request, pk):
     
     return render(request, "base/room.html", context)
 
+@login_required(login_url="login-user")
 def createRoom(request):
     """ Views for the create room form """
     form = RoomForm()
@@ -43,11 +81,15 @@ def createRoom(request):
     context = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login-user")
 def updateRoom(request, pk):
     """ Update room for a particular user """
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to edit this!")
+    
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -57,9 +99,14 @@ def updateRoom(request, pk):
     context = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login-user")
 def deleteRoom(request, pk):
     """ Delete a particular room """
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to delete this!")
+
     if request.method == "POST":
         room.delete()
         return  redirect('home')
